refactor: delete meanless alias
