+++ src/todofile/v2.ts
@@ -1,7 +1,7 @@
 import { relative } from "pathe";
 import typia from "typia";
 
-import type { TodoModuleHandler } from "./types";
+import type { TodoModuleHandler } from "./index";
 
 import { isNonEmptyString } from "../utils/string";
